Export range and position comparators.
Also rename Position to SourcePosition and fixup some imports.
